#119 refactor: Rename these hooks so the files match the export
