+++ services/ui/src/hooks/useGetDevices.ts
